refact(entrypoint): move Analyzer class to lib entrypoint and add tests (#3)
